back/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
dnspython
pydantic
qrcode
pillow
